# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Contrato público do BFF NEXORA para dados operacionais e inteligência.
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  code: string;
  message: string;
}

export type EnvelopeData = { [key: string]: unknown };

export type EnvelopeSource = typeof EnvelopeSource[keyof typeof EnvelopeSource];


export const EnvelopeSource = {
  live: 'live',
} as const;

export interface Envelope {
  data: EnvelopeData;
  source: EnvelopeSource;
  provider?: string;
  lastUpdated: string;
}

export type MetricTrend = typeof MetricTrend[keyof typeof MetricTrend];


export const MetricTrend = {
  up: 'up',
  down: 'down',
  neutral: 'neutral',
} as const;

export interface Metric {
  label: string;
  value: string;
  change: string;
  trend?: MetricTrend;
  note: string;
}

export type OverviewCompany = {[key: string]: string};

export type OverviewRevenueItem = { [key: string]: unknown };

export type OverviewSummary = { [key: string]: unknown };

export type OverviewReceivables = { [key: string]: unknown };

export type OverviewStockCritical = { [key: string]: unknown };

export type OverviewAlertsItem = { [key: string]: unknown };

export interface Overview {
  company: OverviewCompany;
  metrics: Metric[];
  revenue: OverviewRevenueItem[];
  summary: OverviewSummary;
  receivables: OverviewReceivables;
  stockCritical: OverviewStockCritical;
  alerts: OverviewAlertsItem[];
}

export interface Customer {
  id: string;
  name: string;
  initials: string;
  status: string;
  opportunity: number;
  next: string;
  since?: string;
}

export interface CustomerInput {
  name: string;
  email?: string;
}

export interface Customers {
  items: Customer[];
  total: number;
}

export interface Sales { [key: string]: unknown }

export interface Finance { [key: string]: unknown }

export interface Stock { [key: string]: unknown }

export interface Suppliers { [key: string]: unknown }

export interface Alerts { [key: string]: unknown }

export interface Alert { [key: string]: unknown }

export interface Approval { [key: string]: unknown }

export interface Approvals { [key: string]: unknown }

export type DecisionInputDecision = typeof DecisionInputDecision[keyof typeof DecisionInputDecision];


export const DecisionInputDecision = {
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface DecisionInput {
  decision: DecisionInputDecision;
  reason?: string;
}

export type QuestionInputContext = { [key: string]: unknown };

export interface QuestionInput {
  /** @minLength 1 */
  question: string;
  context?: QuestionInputContext;
}

export type IntelligenceMetadata = { [key: string]: unknown };

export interface Intelligence {
  answer: string;
  metadata?: IntelligenceMetadata;
}

export type OverviewEnvelope = Envelope;

export type CustomersEnvelope = Envelope;

export type CustomerEnvelope = Envelope;

export type SalesEnvelope = Envelope;

export type FinanceEnvelope = Envelope;

export type StockEnvelope = Envelope;

export type SuppliersEnvelope = Envelope;

export type AlertsEnvelope = Envelope;

export type AlertEnvelope = Envelope;

export type ApprovalsEnvelope = Envelope;

export type ApprovalEnvelope = Envelope;

export type IntelligenceEnvelope = Envelope;

/**
 * Integração privada indisponível
 */
export type UnavailableResponse = Error;

/**
 * Dados reais do backend NEXORA
 */
export type LiveOverviewResponse = OverviewEnvelope;

/**
 * Clientes reais
 */
export type LiveCustomersResponse = CustomersEnvelope;

/**
 * Cliente real
 */
export type LiveCustomerResponse = CustomerEnvelope;

/**
 * Vendas reais
 */
export type LiveSalesResponse = SalesEnvelope;

/**
 * Finanças reais
 */
export type LiveFinanceResponse = FinanceEnvelope;

/**
 * Stock real
 */
export type LiveStockResponse = StockEnvelope;

/**
 * Fornecedores reais
 */
export type LiveSuppliersResponse = SuppliersEnvelope;

/**
 * Alertas reais
 */
export type LiveAlertsResponse = AlertsEnvelope;

/**
 * Alerta atualizado
 */
export type LiveAlertResponse = AlertEnvelope;

/**
 * Aprovações reais
 */
export type LiveApprovalsResponse = ApprovalsEnvelope;

/**
 * Aprovação atualizada
 */
export type LiveApprovalResponse = ApprovalEnvelope;

/**
 * Resposta do Orquestrador Central de IA
 */
export type LiveIntelligenceResponse = IntelligenceEnvelope;

export type SearchParameter = string;

export type StatusParameter = string;

export type PeriodParameter = string;

export type OpenOnlyParameter = boolean;

export type FinanceTabParameter = string;

export type ListCustomersParams = {
search?: SearchParameter;
status?: StatusParameter;
};

export type GetSalesParams = {
period?: PeriodParameter;
openOnly?: OpenOnlyParameter;
};

export type GetFinanceParams = {
period?: PeriodParameter;
tab?: FinanceTabParameter;
};