# lib/api-zod/src/generated/types/approval.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Contrato público do BFF NEXORA para dados operacionais e inteligência.
 * OpenAPI spec version: 0.2.0
 */

export interface Approval { [key: string]: unknown }