# lib/api-zod/src/generated/types/envelope.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Contrato público do BFF NEXORA para dados operacionais e inteligência.
 * OpenAPI spec version: 0.2.0
 */
import type { EnvelopeData } from './envelopeData';
import type { EnvelopeSource } from './envelopeSource';

export interface Envelope {
  data: EnvelopeData;
  source: EnvelopeSource;
  provider?: string;
  lastUpdated: Date;
}