# lib/api-zod/src/generated/types/intelligence.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Contrato público do BFF NEXORA para dados operacionais e inteligência.
 * OpenAPI spec version: 0.2.0
 */
import type { IntelligenceMetadata } from './intelligenceMetadata';

export interface Intelligence {
  answer: string;
  metadata?: IntelligenceMetadata;
}