# lib/api-zod/src/generated/types/intelligenceEnvelope.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Contrato público do BFF NEXORA para dados operacionais e inteligência.
 * OpenAPI spec version: 0.2.0
 */
import type { Envelope } from './envelope';

export type IntelligenceEnvelope = Envelope;