# lib/api-zod/src/generated/types/metric.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Contrato público do BFF NEXORA para dados operacionais e inteligência.
 * OpenAPI spec version: 0.2.0
 */
import type { MetricTrend } from './metricTrend';

export interface Metric {
  label: string;
  value: string;
  change: string;
  trend?: MetricTrend;
  note: string;
}